Refactor logic for generating agent metrics
Exposing the pool name was difficult because we only collected the
agents and they had no concept of pools baked in and adding that concept
in was difficult because of how the code was structured.

This commit refactors the logic for creating the agent availbility
metric to allow the pool name to be collected easily and to make sense
logically.

The logic has effecitvely come up a layer and is now controlled via the
TFSCollector rather than TFS. The logic is also now split into smaller
chunks and the values between them are passed via channels.

At the moment this code leaks GoRountines as we never close the original
`chanAgents` when all the agents have been completed.