Handle errors during pipeline
While scraping agents errors can occur such as being unable to
scrape a pool due to HTTP requests failing. This means the entire
metrics is incomplete. If only 6 out of 7 pools report back this causes
the metrics to be differnt to if all 7 report back creating false
positives.

This commit will "buffer" until all the pools have been scraped and then
make a decision to expose the collected metrics or not.

If a single pool returns an error, no metrics will be returned for this
scrape of the server.